Reworked constraint parser: numbers are passed around as a slice
The previous version esposed a significant drawback of my previous
approach of keeping all digits in a fixed-length array: there is no flag
that all values were set. This is critical to the caret operation where
the latest set digit defines whether it has to range a patch, minor or major
version.

With this approach a slice contains an extra bit of information: how many digits
were provided. Using this lets us implement a caret operator seamlessly.

Signed-off-by: Oleg Sidorov <me@whitebox.io>